Working on the gateway client and service.
Modified the core bus to add the 'clientId' parameter as a round-trip value (like correlation ID) so that the gateway service can keep track of which client the response belongs to.

The gateway service is demo/alpha quality; it can accept messages and forward them to rabbit, and keep track replies. The client is starting to be able to inject messages.

The messaging across the gateway transport is now split into two separate message types, GatewayInboundMsg and GatewayOutboundMsg since they end up carrying different information.